+++ NPS Browser/controllers/DataViewController.swift
@@ -10,7 +10,6 @@ import Cocoa
 
 class DataViewController: NSViewController, ToolbarDelegate {
     
-//    @IBOutlet weak var btnAddBookmark: NSButton!
     @IBOutlet var tsvResultsController: NSArrayController!
     @IBOutlet weak var tableView: NSTableView!
     lazy var windowDelegate: WindowDelegate = Helpers().getWindowDelegate()
@@ -25,7 +24,8 @@ class DataViewController: NSViewController, ToolbarDelegate {
     }
     
     override func viewDidAppear() {
-        self.windowDelegate = Helpers().getWindowDelegate()
+        
+        CoreDataIO().deleteAll()
         
         windowDelegate.startBtnReloadAnimation()
         if (CoreDataIO().recordsAreEmpty()) {
@@ -51,27 +51,13 @@ class DataViewController: NSViewController, ToolbarDelegate {
         }
     }
     
-    @IBAction func toggleBookmark(_ sender: NSButton) {
-        let rowData = Helpers().getRowObjectFromTableRowButton(sender) as Any
-        
-        let bookmark: Bookmark = Helpers().makeBookmark(rowData: rowData)
-        
-        if (sender.state == .on) {
-            Helpers().getSharedAppDelegate().bookmarkManager.addBookmark(bookmark)
-        } else {
-            Helpers().getSharedAppDelegate().bookmarkManager.removeBookmark(bookmark)
-        }
-    }
-    
     func setArrayControllerContent(content: [NSManagedObject]?) {
         tsvResultsController.content = nil
         tsvResultsController.content = content
-        
         windowDelegate.stopBtnReloadAnimation()
+        tableView.reloadData()
     }
 }
-
-
 
 extension NSTableView {
     override open func keyDown(with event: NSEvent) {