+++ NPS Browser/models/DLItem.swift
@@ -13,36 +13,21 @@ class DLItem: NSObject {
     @objc dynamic var title_id          : String?
     @objc dynamic var name              : String?
     @objc dynamic var type              : String?
-    @objc dynamic var url               : URL?
+    @objc dynamic var pkg_direct_link   : URL?
     @objc dynamic var progress = 0.0
     @objc dynamic var zrif              : String?
     @objc dynamic var status            : String?
     @objc dynamic var timeRemaining     : TimeInterval = 0
-    var request: Alamofire.Request?
+    var request                         : Alamofire.Request?
+    var resumeData                      : Data?
     @objc dynamic var destinationURL    : URL?
-    @objc dynamic var isCancelable      : Bool = true
+    @objc dynamic var isCancelable      : Bool = false
     @objc dynamic var isViewable        : Bool = false
     @objc dynamic var isRemovable       : Bool = false
     @objc dynamic var isResumable       : Bool = false
     
     override init() {
         super.init()
-    }
-    
-    func setName(_ name: String) {
-        self.name = name
-    }
-    
-    func setUrl(_ url: URL) {
-        self.url = url
-    }
-    
-    func setZrif(_ zrif: String) {
-        self.zrif = zrif
-    }
-    
-    func setRequest(_ request: Alamofire.Request) {
-        self.request = request
     }
     
     func makeCancelable() {
@@ -62,6 +47,13 @@ class DLItem: NSObject {
     func makeResumable() {
         self.isResumable    = true
         self.isRemovable    = true
+        self.isCancelable   = false
+        self.isViewable     = false
+    }
+    
+    func makeRemovable() {
+        self.isRemovable    = true
+        self.isResumable    = false
         self.isCancelable   = false
         self.isViewable     = false
     }