+++ NPS Browser/models/WindowDelegate.swift
@@ -15,9 +15,5 @@ protocol WindowDelegate {
     func startBtnReloadAnimation()
     func stopBtnReloadAnimation()
     
-    func addToDownloadQueue(data: DLItem)
-    func removeCompletedFromDownloadQueue()
-    func getDownloadQueue() -> [DLItem]
-    
     func getDataController() -> DataViewController
 }