+++ NPS Browser/services/BookmarkManager.swift
@@ -7,48 +7,69 @@
 //
 
 import Cocoa
+import Foundation
 
 class BookmarkManager {
-    
+
     let cd = CoreDataIO()
-    var bookmarkList: [Bookmark] = []
+    var entity: NSEntityDescription
+    var bookmarkButtonIDArray = [String: NSButton]()
     
     init() {
-        
+        self.entity = cd.getEntity(entityName: "Bookmarks")
     }
     
-    func getBookmarkList() -> [Bookmark] {
-        return self.bookmarkList
+    func getBookmarkList() -> [NSManagedObject?] {
+        return cd.getBookmarks()
     }
-
-    func addBookmark(_ bookmark: Bookmark) {
-        bookmarkList.append(bookmark)
+    
+    func getBookmark(title_id: String) -> NSManagedObject? {
+        return cd.getRecordByTitleID(entityName: "Bookmarks", title_id: title_id)
+    }
+    
+    func addBookmarkButtonIDToArray(title_id: String, button: NSButton) {
+        bookmarkButtonIDArray[title_id] = button
+    }
+    
+    func addBookmark(bookmark: Bookmark, item: NSManagedObject, sender: NSButton) {
+        // Store button object id so we can toggle state when bookmark is removed via the popover
+        addBookmarkButtonIDToArray(title_id: bookmark.title_id, button: sender)
+        saveBookmark(bookmark: bookmark, item: item)
     }
     
     func removeBookmark(_ bookmark: Bookmark) {
-        let index = bookmarkList.index(where: { (item) -> Bool in
-            item.title_id == bookmark.title_id
-        })
+        let obj = cd.getRecordByTitleID(entityName: "Bookmarks", title_id: bookmark.title_id)
+        
+        bookmarkButtonIDArray[bookmark.title_id]?.state = .off
+        bookmarkButtonIDArray.remove(at: bookmarkButtonIDArray.index(forKey: bookmark.title_id)!)
 
-        self.bookmarkList.remove(at: index!)
+//        Helpers().getDataController().tableView.re
+        
+        do {
+            cd.getContext().mergePolicy = NSMergePolicyType.mergeByPropertyObjectTrumpMergePolicyType
+            try cd.getContext().delete(obj!)
+        } catch let error as NSError {
+            print("Could not save. \(error), \(error.userInfo)")
+        }
     }
     
-    func saveBookmarks() {
-        let entity = cd.getEntity(entityName: "Bookmarks")
-        bookmarkList.forEach({ item in
-            let bookmark = cd.getObject(entity: entity)
-            bookmark.setValue(item.name, forKey: "name")
-            bookmark.setValue(item.title_id, forKey: "title_id")
-            bookmark.setValue(item.type, forKey: "type")
-            bookmark.setValue(item.zrif, forKey: "zrif")
-            bookmark.setValue(item.url, forKey: "url")
-//            bookmark.setValue(item.refObjectID, forKey: "item")
-
-            do {
-                try cd.getContext().save()
-            } catch let error as NSError {
-                print("Could not save. \(error), \(error.userInfo)")
-            }
-        })
+    func saveBookmark(bookmark: Bookmark, item: NSManagedObject) {
+        let obj = cd.getObject(entity: self.entity)
+        obj.setValue(bookmark.name, forKey: "name")
+        obj.setValue(bookmark.title_id, forKey: "title_id")
+        obj.setValue(bookmark.type, forKey: "type")
+        obj.setValue(bookmark.zrif, forKey: "zrif")
+        obj.setValue(bookmark.pkg_direct_link, forKey: "pkg_direct_link")
+        obj.setValue(item, forKey: "item")
+        
+        do {
+            try cd.getContext().save()
+        } catch let error as NSError {
+            print("Could not save. \(error), \(error.userInfo)")
+        }
+    }
+    
+    func deleteAll() {
+        self.cd.deleteAll()
     }
 }