+++ NPS Browser/services/DownloadManager.swift
@@ -13,7 +13,7 @@ import Alamofire
 class DownloadManager {
     
     var downloadItems: [DLItem] = []
-    let queue = Queuer(name: "DLQueue", maxConcurrentOperationCount: 3, qualityOfService: .default)
+    let queue = Queuer(name: "DLQueue", maxConcurrentOperationCount: SettingsManager().getDownloads().concurrent_downloads, qualityOfService: .default)
     
     init() {}
     
@@ -29,8 +29,8 @@ class DownloadManager {
         }
         
         // store request in same object so we can cancel/pause/resume it later
-        let request = Alamofire.download(data.url!, to: destination)
-        data.setRequest(request)
+        let request = Alamofire.download(data.pkg_direct_link!, to: destination)
+        data.request = request
         
         // add object to downloadItems array
         downloadItems.insert(data, at: 0)
@@ -42,7 +42,8 @@ class DownloadManager {
         
         let dlFileOperation = ConcurrentOperation {
             request.downloadProgress { progress in
-                dlItem.status = "Downloading..."
+                dlItem.status = "Downloading..."        
+                dlItem.makeCancelable()
                 dlItem.progress = (progress.fractionCompleted * 100).rounded()
                 dlItem.timeRemaining = progress.fractionCompleted
             }
@@ -56,21 +57,49 @@ class DownloadManager {
                 response.result.ifFailure {
                     guard let resumeData = response.resumeData else {
                         dlItem.status = "Failed! \(response.error.debugDescription)"
-                        dlItem.isCancelable = false
+                        dlItem.makeRemovable()
                         return
                     }
                     dlItem.status = "Download Cancelled"
-                    dlItem.isResumable = true
+                    dlItem.resumeData = resumeData
+                    dlItem.makeResumable()
                 }
             }
         }
         self.queue.addOperation(dlFileOperation)
     }
+    
+    func resumeDownload(data: DLItem) {
+        let request = Alamofire.download(resumingWith: data.resumeData!)
+        data.request = request
+        
+        request.downloadProgress { progress in
+            data.status = "Downloading..."
+            data.makeCancelable()
+            data.progress = (progress.fractionCompleted * 100).rounded()
+            data.timeRemaining = progress.fractionCompleted
+        }
+            .responseData { response in
+                response.result.ifSuccess {
+                    ExtractionManager(item: data, downloadManager: self).start()
+                }
+                response.result.ifFailure {
+                    guard let resumeData = response.resumeData else {
+                        data.status = "Failed! \(response.error.debugDescription)"
+                        data.makeRemovable()
+                        return
+                    }
+                    data.status = "Download Cancelled"
+                    data.resumeData = resumeData
+                    data.makeResumable()
+                }
+        }
+    }
 
     func removeCompleted() {
         for item in downloadItems {
             if (item.isRemovable) {
-                moveToCompleted(item: item)
+                downloadItems.remove(at: downloadItems.index(of: item)!)
             }
         }
     }