+++ CHANGELOG.md
@@ -10,14 +10,13 @@
 - Completed downloads move to bottom of download list
 - Button added to download list items to allow resuming a cancelled download.
 - Arrow keys can now be used to navigate the data table in the master view
-- Bookmarks may now be added to the Bookmarks window by clicking the star by any entry in the master view
+- Concurrent download amount can be changed in the preferences window.
 
 ### Changed
 - Changed repository name to "NPS-Browser-macOS" to avoid confusion with original NPS Browser.
 - Changed app name to "NPS Browser"
 - Fixed a bug that would crash the app when trying to clear download list witih more than 1 item
-- Fixed the save button in the preferences window, now closes the window after saving
-- Preferences window now cannot be resized or maximized
+- UI tweaks
 - Application terminates when the red close button is pressed instead of just closing the window
 
 ## 1.1 - 2018-06-08